requirements.txt
ccxt
pandas
pandas-ta
python-dotenv
